pulled util_v out of model.
lesson learned: it was not an appropriate place for it, because its
length depended on the reference alignment. the ref align is not part of
the model.

it is now in Three_tax, which isn't perfect, but it is with other things
(Glv's) whose length does depend of the length of the reference
alignment.