Readding in the fantasy baseball functionality.
This required some refactoring of Pplacer_run and Core, but I'm happy with the
results of it. The code is better-organized now than it was. The only thing I'm
slightly annoyed with is the pattern-matching required in the `nextfunc`
implementations. This could be removed, but it would take more work than I
really wanted to spend on it right now.