Comments and minor cleaning. Closes #20.
kr_distance doesn't have a clear way to get faster IMHO. We must keep
things in pres for shuffling, and then we have to change to maps for a
walk along the tree. The "heavy lifting" is done by addition and scaling
of two element arrays. The only thing I can think would be to replace
some of this by c code, but it's not clear to me that this would be
faster given that we are looking at two-element arrays.

I'm not trying to make this faster for the parameter of tree size, just
for the number of placements.