Shift in simulation methodology
Now allowing subtrees to choose overlapping lsetsets. So some taxa can
get thrown out, and some can get chosen multiple times.

Doing this because I would like the number of splits chosen to be
related to the "signal" along that edge. That way, the poisson mean can
be like a branch length.