Adding BIOM support to guppy/rppr.
Anything that reads a placefile is now able to read an OTU table BIOM file with
a tree as if it was a split placefile. Reference packages can also be built by
splitting an OTU table BIOM file into a taxtable and seqinfo file. This allows
taxonomically-annotated analyses to be done on QIIME output.